Updated The Lab.
Brought to nearly full functionality. Player can transform several different items, including the flask which eventually becomes a key, and can use the key to open the door to the next room.

Tweaked "open" action section of main so that Lab leads to an undefined fourth room. This required adding an undefined room creation method in RoomSetup

Known bugs:
-After learning how to use machine, player can use transmorgable and transform them items without entering machine use menu.